requirements.txt
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.3
django-channels-jwt-auth-middleware==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
ez-setup==0.9
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
msgpack==1.0.4
mysql==0.0.3
mysqlclient==2.1.1
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.32
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.6
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.1
tomli==2.0.1
traitlets==5.5.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
uWSGI==2.0.21
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.5.1
